Add discord install

+++ initial-startup/debian/install_discord.sh
@@ -1,3 +1,5 @@
-# install discord
-wget -O discord.deb https://discord.com/api/download?platform=linux&format=deb
-sudo dpkg -i discord.deb
+# install discord if it is not installed
+if ! dpkg -l | grep discord; then
+  wget -O discord.deb https://discord.com/api/download?platform=linux&format=deb
+  sudo dpkg -i discord.deb
+fi
